Pizza Controller & Services Updated

+++ src/main/java/com/pizzaApp/Services/PizzaServices.java
@@ -1,5 +1,7 @@
 package com.pizzaApp.Services;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,7 +12,7 @@ import com.pizzaApp.Repository.PizzaRepository;
 @Service
 public class PizzaServices {
 	@Autowired
-    PizzaRepository pizzaRepository;
+    PizzaRepository pizzaRepository;	
     public boolean add(Pizza pizza) {
         Pizza piz=pizzaRepository.findBypizzaName(pizza.getPizzaName());
         if(piz==null){
@@ -25,7 +27,28 @@ public class PizzaServices {
         
     	public void deletePizza(int pizzaId) 
     	{
-    		//Pizza pizza=PizzaRepository.findById(pizzaId);
-    		pizzaRepository.deleteById(pizzaId);      
+    		if(pizzaRepository.findById(pizzaId).isPresent())
+    		{
+    			pizzaRepository.deleteById(pizzaId);
+    		}
     	}
+
+		public void updatePizza(Pizza pizza) 
+		{
+			if(pizzaRepository.findById(pizza.getPizzaId()).isPresent())
+    		{
+    			Pizza p=pizzaRepository.findById(pizza.getPizzaId()).get();
+    			p.setPizzaCost(pizza.getPizzaCost());
+    			p.setPizzaDescription(pizza.getPizzaDescription());
+    			p.setPizzaName(pizza.getPizzaName());
+    			p.setPizzaType(pizza.getPizzaType());
+    			pizzaRepository.save(p);
+    		}
+		}
+
+		public List viewPizzaList() 
+		{
+			return pizzaRepository.findAll();
+		}
+    	
     }
